requirements.txt
PyYAML~=6.0.2
ezdxf~=1.3.3
numpy~=2.0.1
opencv-python~=4.10.0.84
pymysql~=1.1.1
fastapi~=0.115.0
pydantic~=2.9.2
